Publish known `system_info` data from device to message bus
When hardware information from the device the device through the
`/system_info` endpoint, make a best effort to parse the `product`
from the json and publish a message to the bus containing that data.

Currently we already publish a msg with that information, this commit
augments that message with an optional `product` attribute.

Signed-off-by: Simão Mata <simao.mata@here.com>