Remove param to filter not activated devices.
It seems this is not a very useful use-case after all. Now we want to
show the devices that have not connected in the last X hours (OTA-4035).
This `activated` param has not been use in the end and there is no plan
to use it, so I think it's better to remove it and clean-up before we
introduce a new param to search by not-seen in the last hours.

Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>