Take over the "blacklisting" feature from ota-core.
The now-deprecated ota-core service serves a "blacklisting"
feature. The quotes are because it's not actually
blacklisting anything. It is used to count the number of
devices that have a particular package installed. This are
the installed packages reported by aktualizr, e.g. 'nano' for
a linux distribution, not to be confused with the images we
can install through our system.

We move that functionality here to discontinue the ota-core
service.

Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>