Fix sota-core blacklist migration (reprise!).
The symptoms of the failing migration was that it was migrating the
first ~30 records, and then finishing successfully (!).

The issue was that `runForeach` takes a function that returns a `Unit`,
but I was giving it a function that returns a `Future`. The `Future`
was casted to a `Unit` and the migration didn't wait for the `Future`
to complete, finishing once the stream source was over.

To fix it, one should use `mapAsyncUnordered` or `foldAsync` when the
function we pass returns a `Future` and we need to wait for it.

Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>