Create device tags soon to be used in smart groups.
There is a new feature that consists on using custom device fields
to use for smart group. The idea is that each environment/namespace
member can "tag" devices with "custom device fields" that can later
be used to create smart groups.

The first iteration of the feature only requires to be able to see
a disabled device tag. So for now we only need to be able to create
and fetch device tags. A route will be needed in web-app to enable
the FE to fetch the tags. They can't use this API directly because
it's at the namespace level and the FE doesn't have direct access to
the namespace.

I decided to use an incrementable integer as a tag id to keep things
simple when I need to make changes to the smart groups grammar in the
next days, but depending on how things go I might need to tweak the DB
schema.

Also, notice that business are calling these "device custom fields",
but IMHO that sounds confusing and rather inaccurate, and "device tags"
sounds more fitting to me. But I'm open to change the naming if there
is a reasonable reason for it.

Relates-to: OTA-4801
Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>