Validate device tag names as specified.
We specified that "Custom device fields names: alphanumeric characters,
space, underscore, max length 15 characters, min length 1 character",
cfr. related ticket.

The original idea behind this was to be very restrictive with the names
to use them as primary key. We eventually discarded that idea, but it is
still reasonable to be restrictive here, and relax the constraints in the
future if it ever becomes necessary.

Relates-to: OTA-4803
Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>