Fail gracefully with 400 when the csv headers are invalid.
Before the import was crashing with a 500 error. Now it handles
it more gracefully with a 400 code.

Also, the `tagId`s should be allowed to include underscores. I
thought that was the case, since the regex `\w` normally includes
alphanumeric characters as well as underscores. But for Java it
doesn't include underscores [1], so I added it explicitly.

[1] https://docs.oracle.com/javase/tutorial/essential/regex/pre_char_classes.html

Relates-to: OTA-4802
Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>