Removes devices from smart groups when tagging them via CSV.
There's a bug where devices will not be removed from a smart group
if we upload a new CSV and the devices no longer match the expression
of the group.

To fix it, and given that uploading CSV is a rare operation, the surest
way I see is to wipe all the groups for the devices and then add them
to the necessary smart groups.

Relates-to: OTA-4805
Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>