Fix updating group expression after tag renaming.
I had forgotten the most important check on one test, and thus
didn't realise about this misstep earlier. But the query to
update the group expressions after renaming a device tag was
slightly off.

Also, I have to rectify what I said when I introduced this query:
it is actually important to update the expression, because it
is shown by the UI.

Relates-to: OTA-4861
Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>