Add endpoint to delete device custom fields.
This is complex, because when removing a tag/custom field, we need
to remove it from the expressions of the smart groups that use that
tag. And in turn that requires adjusting the group members. Indeed,
after removing the tag from the expression, some devices might not
match the expression anymore, and some others might start to match it.

If there is at least one group that uses that tag, and only that tag,
the request will fail. That's because removing the tag would lead to an
empty group expression. Otherwise, for each smart group that uses that
tag, we calculate the resulting expression after removing the tag, we remove
all the members, and for all the devices in the namespace we check and
potentially add them to the group.

Relates-to: OTA-4862
Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>